cli: move backup/restore under debug
And get rid of the `experimental` build tag, this file was not checked by the CI
and was too easy to break. Three files (one for the code, plus one each for the
build tag and its inverse) would also have worked, but debug is just as good a
place to hide an in-progress feature and less hassle.
