+++ cli/cli.go
@@ -87,7 +87,6 @@ func init() {
 		versionCmd,
 		debugCmd,
 	)
-	cockroachCmd.AddCommand(backupCmds...)
 }
 
 // Run ...
