+++ cli/debug.go
@@ -769,6 +769,8 @@ var debugCmds = []*cobra.Command{
 	kvCmd,
 	rangeCmd,
 	debugEnvCmd,
+	backupCmd,
+	restoreCmd,
 }
 
 var debugCmd = &cobra.Command{
